servicio implementacion de Administrador

+++ src/main/java/org/esfe/BeatySaly/repositorios/IAdministradorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IAdministradorRepository extends JpaRepository<Administrador, Long> {
+public interface IAdministradorRepository extends JpaRepository<Administrador, Integer> {
     Administrador findByUsername(String username);
 }
